+++ lib/src/appbars/appbar_back.dart
@@ -27,7 +27,11 @@ PreferredSizeWidget AppBarBack({
     toolbarHeight: 65,
     backgroundColor: effectiveColor,
     elevation: 0,
-    title: TextSubtitle(title, color: effectiveTextColor),
+    title: TextBody(
+      title,
+      color: effectiveTextColor,
+      fontWeight: WEIGHT_BOLD,
+    ),
     leading: IconButton(
       splashColor: COLOR_ACCENT.withValues(alpha: 0.2),
       highlightColor: COLOR_ACCENT.withValues(alpha: 0.2),